refactor: extract news uploading into VoicesUploader
Moves uploadNews() and its news-only helpers (createImage,
queueNewsRetry, uploadNewsActivities) out of UploadManager into a
VoicesUploader, mirroring TeamsUploader/PhotoUploader: one domain per
uploader, with the bulk_docs POST still going through BulkDocsUploader.

UploadManager.uploadNews() stays as a delegate, like uploadTeams() and
uploadAchievement(), so AutoSyncWorker and UserDataWorker are unchanged.
The five constructor dependencies only the news path used (context, gson,
retryQueue, voicesRepository, timeProvider) and their imports go with it.

processInBatches was file-private in both UploadManager and TeamsUploader;
now that both callers sit in services/upload it moves to UploadConstants.

Adds VoicesUploaderTest (moved mimeType/header case plus coverage for
marking accepted docs uploaded, queueing a retry on a failed bulk request,
and uploading news activities after the batch); UploadManagerTest gains a
delegation test in place of the moved one.
