+++ app/src/main/java/org/ole/planet/myplanet/services/upload/UploadConstants.kt
@@ -3,3 +3,7 @@ package org.ole.planet.myplanet.services.upload
 object UploadConstants {
     const val BATCH_SIZE = 50
 }
+
+internal inline fun <T> Iterable<T>.processInBatches(action: (List<T>) -> Unit) {
+    chunked(UploadConstants.BATCH_SIZE).forEach(action)
+}
